requirements.txt
# Requirements

configargparse==1.0
imageio==2.6.1
numpy==1.18.1
pillow==7.0.0
torch==1.4.0
pyyaml==5.3
scikit-image==0.16.2
scipy==1.4.1
tensorflow-gpu==1.14.0
torchvision==0.5.0
tqdm==4.42.1
